requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.0
Pillow==9.0.1
ipython==8.5.0
# google-cloud-vision==3.0.1
scikit-learn==1.3.2
pandas==2.2.2
matplotlib==3.9.0
openpyxl==3.1.3


#Para instalar: pip install -r requirements.txt
